+++ Tests/RSocketReactiveSwiftTests/RSocketReactiveSwiftTests.swift
@@ -1,98 +1,33 @@
+/*
+ * Copyright 2015-present the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import XCTest
-import NIO
 import ReactiveSwift
-@testable import RSocketCore
+import RSocketCore
 import RSocketTestUtilities
 @testable import RSocketReactiveSwift
-
-struct TestDemultiplexer {
-    let router: DemultiplexerRouter
-    let requester: Requester
-    let responder: Responder
-    
-    internal init(connectionSide: ConnectionRole, requester: Requester, responder: Responder) {
-        self.router = .init(connectionSide: connectionSide)
-        self.requester = requester
-        self.responder = responder
-    }
-    
-    func receiveFrame(frame: Frame) {
-        let route = router.route(for: frame.header.streamId, type: frame.header.type)
-        if route.contains(.connection) {
-            XCTFail("connection message not expected \(frame)")
-        }
-        if route.contains(.requester) {
-            requester.receiveInbound(frame: frame)
-        }
-        if route.contains(.responder) {
-            responder.receiveInbound(frame: frame)
-        }
-    }
-}
-
-extension TestDemultiplexer {
-    static func pipe(
-        serverResponder: RSocketCore.RSocket,
-        clientResponder: RSocketCore.RSocket
-    ) -> (eventLoop: EmbeddedEventLoop, server: TestDemultiplexer, client: TestDemultiplexer) {
-        var client: TestDemultiplexer!
-        let eventLoop = EmbeddedEventLoop()
-        let server = TestDemultiplexer(
-            connectionSide: .server,
-            requester: .init(streamIdGenerator: .server, eventLoop: eventLoop, sendFrame: { frame in
-                client.receiveFrame(frame: frame)
-            }),
-            responder: .init(responderSocket: serverResponder, eventLoop: eventLoop, sendFrame: { frame in
-                client.receiveFrame(frame: frame)
-            }))
-        client = TestDemultiplexer(
-            connectionSide: .client,
-            requester: .init(streamIdGenerator: .client, eventLoop: eventLoop, sendFrame: { frame in
-                server.receiveFrame(frame: frame)
-            }),
-            responder: .init(responderSocket: clientResponder, eventLoop: eventLoop, sendFrame: { frame in
-                server.receiveFrame(frame: frame)
-            }))
-        return (eventLoop, server, client!)
-    }
-}
 
 func setup(
     server: RSocketReactiveSwift.RSocket? = nil,
     client: RSocketReactiveSwift.RSocket? = nil
-) -> (eventLoop: EmbeddedEventLoop, server: TestDemultiplexer, client: TestDemultiplexer) {
-    let (eventLoop, server, client) = TestDemultiplexer.pipe(
+) -> (server: TestDemultiplexer, client: TestDemultiplexer) {
+    let (server, client) = TestDemultiplexer.pipe(
         serverResponder: ResponderAdapter(responder: server),
         clientResponder: ResponderAdapter(responder: client))
-    return (eventLoop, server, client)
-}
-
-final class TestRSocket: RSocketReactiveSwift.RSocket {
-    var metadataPushCallback: (Data) -> () = { _ in }
-    var fireAndForgetCallback: (Payload) -> () = { _ in }
-    var requestResponseCallback: (Payload) -> SignalProducer<Payload, Swift.Error> = { _ in .never }
-    var requestStreamCallback: (Payload) -> SignalProducer<Payload, Swift.Error> = { _ in .never }
-    var requestChannelCallback: (Payload, SignalProducer<Payload, Swift.Error>?) -> SignalProducer<Payload, Swift.Error> = { _, _ in .never }
-    
-    internal init(
-        metadataPush: @escaping (Data) -> () = { _ in },
-        fireAndForget: @escaping (Payload) -> () = { _ in },
-        requestResponse: @escaping (Payload) -> SignalProducer<Payload, Swift.Error> = { _ in .never },
-        requestStream: @escaping (Payload) -> SignalProducer<Payload, Swift.Error> = { _ in .never },
-        requestChannel: @escaping (Payload, SignalProducer<Payload, Swift.Error>?) -> SignalProducer<Payload, Swift.Error> = { _, _ in .never }
-    ) {
-        self.metadataPushCallback = metadataPush
-        self.fireAndForgetCallback = fireAndForget
-        self.requestResponseCallback = requestResponse
-        self.requestStreamCallback = requestStream
-        self.requestChannelCallback = requestChannel
-    }
-    
-    func metadataPush(metadata: Data) { metadataPushCallback(metadata) }
-    func fireAndForget(payload: Payload) { fireAndForgetCallback(payload) }
-    func requestResponse(payload: Payload) -> SignalProducer<Payload, Swift.Error> { requestResponseCallback(payload) }
-    func requestStream(payload: Payload) -> SignalProducer<Payload, Swift.Error> { requestStreamCallback(payload) }
-    func requestChannel(payload: Payload, payloadProducer: SignalProducer<Payload, Swift.Error>?) -> SignalProducer<Payload, Swift.Error> { requestChannelCallback(payload, payloadProducer) }
+    return (server, client)
 }
 
 final class RSocketReactiveSwiftTests: XCTestCase {
@@ -103,7 +38,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
             didReceiveRequest.fulfill()
             XCTAssertEqual(data, metadata)
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         client.requester.rSocket.metadataPush(metadata: metadata)
         self.wait(for: [didReceiveRequest], timeout: 0.1)
     }
@@ -113,7 +48,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
             didReceiveRequest.fulfill()
             XCTAssertEqual(payload, "Hello World")
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         client.requester.rSocket.fireAndForget(payload: "Hello World")
         self.wait(for: [didReceiveRequest], timeout: 0.1)
     }
@@ -129,7 +64,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
                 observer.sendCompleted()
             }
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         let disposable = client.requester.rSocket.requestResponse(payload: "Hello World").startWithSignal { signal, _ in
             signal.flatMapError({ error in
                 XCTFail("\(error)")
@@ -160,7 +95,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
                 observer.sendCompleted()
             }
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         let disposable = client.requester.rSocket.requestStream(payload: "Hello World").startWithSignal { signal, _ in
             signal.flatMapError({ error in
                 XCTFail("\(error)")
@@ -209,7 +144,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
                 observer.sendCompleted()
             }
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         let requestSocket = client.requester.rSocket
         let disposable = requestSocket.requestChannel(payload: "Hello Responder", payloadProducer: .init({ observer, _ in
             requesterDidSendChannelMessages.fulfill()
@@ -254,7 +189,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
                 }
             }
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         let disposable = client.requester.rSocket.requestResponse(payload: "Hello World").startWithSignal { signal, _ -> Disposable? in
             didStartRequestSignal.fulfill()
             return signal.flatMapError({ error -> Signal<Payload, Never> in
@@ -283,7 +218,7 @@ final class RSocketReactiveSwiftTests: XCTestCase {
                 }
             }
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         let disposable = client.requester.rSocket.requestStream(payload: "Hello World").startWithSignal { signal, _ -> Disposable? in
             didStartRequestSignal.fulfill()
             return signal.flatMapError({ error -> Signal<Payload, Never> in
@@ -297,28 +232,9 @@ final class RSocketReactiveSwiftTests: XCTestCase {
         disposable?.dispose()
         self.wait(for: [didReceiveRequest, didEndLifetimeOnResponder], timeout: 0.1)
     }
-//    func testSignalLifetime() {
-//        let signalProducerLifetimeEnded = expectation(description: "signal producer lifetime ended")
-//        var observerStrongRef: Signal<Void, Never>.Observer?
-//        let signalProducer = SignalProducer<Void, Never> { observer, lifetime in
-//            observerStrongRef = observer
-//            lifetime.observeEnded {
-//                signalProducerLifetimeEnded.fulfill()
-//            }
-//        }
-//        let signalInterrupted = self.expectation(description: "signal interrupted")
-//        let disposable = signalProducer.startWithSignal { signal, _ in
-//            signal.observeInterrupted {
-//                signalInterrupted.fulfill()
-//            }
-//        }
-//        disposable?.dispose()
-//        wait(for: [signalProducerLifetimeEnded, signalInterrupted], timeout: 0.1)
-//    }
     func testRequestChannelCancellation() {
         let didReceiveRequestChannel = expectation(description: "did receive request channel")
-        let requesterDidSendChannelMessages = expectation(description: "requester did send channel messages")
-        let responderDidSendChannelMessages = expectation(description: "responder did send channel messages")
+        let responderDidStartSenderProducer = expectation(description: "responder did start sender producer")
         let responderDidStartListeningChannelMessages = expectation(description: "responder did start listening to channel messages")
         let responderProducerLifetimeEnded = expectation(description: "responder producer lifetime ended")
         
@@ -337,19 +253,20 @@ final class RSocketReactiveSwiftTests: XCTestCase {
             }
 
             return SignalProducer { observer, lifetime in
-                responderDidSendChannelMessages.fulfill()
+                responderDidStartSenderProducer.fulfill()
                 lifetime.observeEnded {
                     _ = observer
                     responderProducerLifetimeEnded.fulfill()
                 }
             }
         })
-        let (_, _, client) = setup(server: serverResponder)
+        let (_, client) = setup(server: serverResponder)
         let requestSocket = client.requester.rSocket
         let requesterDidStartListeningChannelMessages = expectation(description: "responder did start listening to channel messages")
         let payloadProducerLifetimeEnded = expectation(description: "payload producer lifetime ended")
+        let requesterDidStartPayloadProducer = expectation(description: "requester did start payload producer")
         let disposable = requestSocket.requestChannel(payload: "Hello", payloadProducer: .init({ observer, lifetime in
-            requesterDidSendChannelMessages.fulfill()
+            requesterDidStartPayloadProducer.fulfill()
             lifetime.observeEnded {
                 _ = observer
                 payloadProducerLifetimeEnded.fulfill()
@@ -365,9 +282,9 @@ final class RSocketReactiveSwiftTests: XCTestCase {
         }
         self.wait(for: [
             didReceiveRequestChannel,
-            requesterDidSendChannelMessages,
+            requesterDidStartPayloadProducer,
             responderDidStartListeningChannelMessages,
-            responderDidSendChannelMessages,
+            responderDidStartSenderProducer,
             requesterDidStartListeningChannelMessages,
         ], timeout: 0.1)
         disposable?.dispose()
